refactor: move all store configuration into providers
